requirements.txt
openai-agents>=0.2.8
fastapi>=0.112
uvicorn>=0.30
pydantic>=2.7
python-dotenv>=1.0
httpx>=0.27
google-api-python-client>=2.143
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
pytest>=8.2
